v1.3
Added card output to the build options, including buttons to output cards, and the ability to split those over elem rows in a menu. Also added a list (l) argument and sfxlist argument to the getrepeating function to accommodate this. The sfxlist argument can list the suffixes for the sub-attributes to include in the card output for a given repeating element. The list argument will disambiguate the repeating element provided, including testing against the current value of the naming element in the group (meaning it will catch the spell, 'Magic Missile' in the spells repeating section if you have an entry whose sfxn has a current value of 'Magic Missile').